+++ changelog.md
@@ -5,6 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Fix conflicting detection heuristics for Linux distributions. Closes #72.
+
 ## [1.3.0 (2024-09-11)](https://github.com/kdeldycke/extra-platforms/compare/v1.2.1...v1.3.0)
 
 - Add detection of all versions of macOS and Windows. Closes #55.